infra/requirements.txt
aws-cdk-lib==2.138.1
constructs>=10.3.0,<11.0.0
